Start moving attention values out of the atom
